Optimize performance of Commerce blocks (#44)
* Optimize custom PLP code

* Load Commerce code early

* Add custom PLP block to LCP block list

* Optimize PLP widget

* Fix PLP CLS on mobile

* Add store header from config

* Remove obsolete product-enrichment block

* Add product name to document title

* Fix bug in performMonolithGraphQLQuery for API Mesh
