+++ blocks/product-list-page/product-list-page.css
@@ -1,5 +1,15 @@
 /* stylelint-disable selector-class-pattern */
+.product-list-page {
+    min-height: 700px;
+}
+
 .product-list-page .ds-sdk-product-item__image img {
     aspect-ratio: 185 / 230;
     object-fit: contain;
+}
+
+@media (width >= 750px) {
+    .product-list-page {
+        min-height: auto;
+    }
 }
