+++ head.html
@@ -1,4 +1,6 @@
 <meta name="viewport" content="width=device-width, initial-scale=1"/>
 <script src="/scripts/aem.js" type="module"></script>
 <script src="/scripts/scripts.js" type="module"></script>
+<script src="/scripts/configs.js" type="module"></script>
+<script src="/scripts/commerce.js" type="module"></script>
 <link rel="stylesheet" href="/styles/styles.css"/>
